third_party/WebKit/LayoutTests/http/tests/cachestorage/serviceworker/cache-match-expected.txt

This is a testharness.js-based test.
PASS Cache.match and Cache.matchAll 
FAIL Cache.matchAll with no matching entries promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with no matching entries 
FAIL Cache.matchAll with URL promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with URL 
FAIL Cache.matchAll with Request promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with Request 
FAIL Cache.matchAll with new Request promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with new Request 
FAIL Cache.matchAll with ignoreSearch option (request with no search parameters) promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with ignoreSearch option (request with no search parameters) 
FAIL Cache.matchAll with ignoreSearch option (request with search parameter) promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with ignoreSearch option (request with search parameter) 
FAIL Cache.matchAll with URL containing fragment promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with URL containing fragment 
FAIL Cache.matchAll with string fragment "http" as query promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with string fragment "http" as query 
FAIL Cache.matchAll with URL containing credentials promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with URL containing credentials 
FAIL Cache.matchAll with responses containing "Vary" header promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
FAIL Cache.match with responses containing "Vary" header assert_true: Cache.match should honor "Vary" header. expected true got false
FAIL Cache.matchAll with "ignoreVary" parameter promise_test: Unhandled rejection with value: object "NotSupportedError: Method is not implemented."
PASS Cache.match with Request and Response objects with different URLs 
PASS Cache.match invoked multiple times for the same Request/Response 
FAIL Cache.match with POST Request assert_equals: Cache.match should not find a match expected (undefined) undefined but got (object) object "[object Response]"
PASS Cache.match with a non-2xx Response 
PASS Cache.match with a network error Response 
Harness: the test ran to completion.
